refactor: Add folder to add typescript types
