+++ src/Context/Types/QuizContextType.tsx
